refactor: event types
