+++ code/cell.module/src/Module/Module.register.ts
@@ -86,7 +86,7 @@ export function registerChild(args: {
 }
 
 /**
- * Listener for "Module/register"events.
+ * Listener for "Module/register" events.
  */
 export function listen<T extends P>(bus: B, module: t.IModule<T>) {
   rx.payload<t.IModuleRegisterEvent>(bus.event$, 'Module/register')
