+++ code/cell.types/src/common/types.ts
@@ -3,7 +3,17 @@ export { Observable, Subject } from 'rxjs';
 export * from '@platform/types';
 export * from '@platform/fs.types';
 export * from '@platform/http.types';
-export * from '@platform/state.types';
+
+export {
+  INode,
+  NodeIdentifier,
+  ITreeNode,
+  ITreeStatePatched,
+  ITreeState,
+  ITreeStateChanged,
+  TreeIdentity,
+  TreeQuery,
+} from '@platform/state.types';
 
 export { TreeNodeIcon } from '@platform/ui.tree.types';
 export { IMemoryCache } from '@platform/cache/lib/types';
