+++ code/state.types/src/types.ts
@@ -1,6 +1,7 @@
 export * from './types.Node';
 export * from './types.Patch';
 export * from './types.StateObject';
+export * from './types.StateObject.events';
 export * from './types.Store';
 export * from './types.TreeIdentity';
 export * from './types.TreeQuery';
